* develop:
  Added multiband support to RasterSourceRelation.
  Provided means of specifying tile size via RasterSourceDataSource.
  Created a less verbose way of getting a schema from an implicit CatalystSerializer.
  Initial skeleton of RasterSourceRelation.
  Renamed `bounds_geometry` to `extent_geometry`.
  Added `st_extent`.
  Renamed `GeometryToBounds` to  `GeometryToExtent`.
  Test fixes associated with changing `bounds` to `geometry`.
  Renamed `withBounds` to `withGeometry`. Added `withExtent`.
  Moved JP2 test to integration test configuration.
  Upgraded to latest geotrellis-contrib-gdal.
  Experiment in delayed evaluation/parsing of CRS specification.
  Added better caching of header info in RasterSource.
  CircleCI build configuration (+5 squashed commits) Squashed commits: [1dd3d111] remove circleci files [cc2a261e] circleci fixes [9b5bcaef] add dockerhub auth for circlci container [87161ea0] add s22s to circleci container name [c73780c5] add gdal to circleci container
  Misc refinements and addition of JP2 test.
  Replaced all but HDFS RasterSource with geotrellis-contrib versions.
  Remove read callbacks due to incompatibility with GeoTrellis RasterSource.
  Incremental reimplementation of RasterSource reading to depend on geotrellis-contrib version. Linting fixes/cleanup.
  Removed `timestamp` from `RasterSource`.
  Removed use of Either in RasterSource and always return MBT. (+1 squashed commit) Build and test tweaks.

# Conflicts:
#	core/src/main/scala/astraea/spark/rasterframes/expressions/TileAssembler.scala

+++ datasource/src/main/scala/astraea/spark/rasterframes/datasource/geotrellis/package.scala
@@ -21,7 +21,7 @@ package astraea.spark.rasterframes.datasource
 import java.net.URI
 
 import _root_.geotrellis.spark.LayerId
-import astraea.spark.rasterframes.datasource.geotrellis.DefaultSource._
+import astraea.spark.rasterframes.datasource.geotrellis.GeoTrellisLayerDataSource._
 import astraea.spark.rasterframes.{RasterFrame, _}
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions.col
@@ -79,12 +79,12 @@ package object geotrellis extends DataSourceOptions {
   implicit class GeoTrellisReaderWithRF(val reader: GeoTrellisRasterFrameReader) {
     def withTileSubdivisions(divs: Int): GeoTrellisRasterFrameReader =
       tag[GeoTrellisRasterFrameReaderTag][DataFrameReader](
-        reader.option(TILE_SUBDIVISIONS_PARAM, divs)
+        reader.option(TILE_SUBDIVISIONS_PARAM, divs.toLong)
       )
 
     def withNumPartitions(partitions: Int): GeoTrellisRasterFrameReader =
       tag[GeoTrellisRasterFrameReaderTag][DataFrameReader](
-        reader.option(NUM_PARTITIONS_PARAM, partitions)
+        reader.option(NUM_PARTITIONS_PARAM, partitions.toLong)
       )
 
     def loadRF(uri: URI, id: LayerId): RasterFrame =
